BUG#57953 my_load_defaults return junk argument ----args-separator---- to caller
After fix of bug#25192, load_defaults() will add an args separator
to distinguish options loaded from configure files from that provided
in the command line. One problem of this is that the args separator
would be added no matter the application need it or not.

Fixed the problem by adding an option:
  bool my_getopt_use_args_separator;
to control whether the separator will be added or not. And also
added functions:
  bool my_getopt_is_args_separator(const char* arg);
to check if the argument is the separator or not.
--BZR--
revision-id: hezx@greatopensource.com-20110117074437-v3icyuwsocqj0j6c
property-branch-nick: 5.5
testament3-sha1: b3535eb1f3cb5dd0fad11ed7d43f6076253269f9