Bug #55284 diagnostics: Enable UNIV_DEBUG_FILE_ACCESSES by UNIV_DEBUG
It was the enabling of UNIV_DEBUG_FILE_ACCESSES that caught Bug #55284
in the first place. This is a very light piece of of debug code, and
there really is no reason why it is not enabled in all debug builds.

rb://551 approved by Jimmy Yang
--BZR--
revision-id: marko.makela@oracle.com-20110202120512-xayv49lo50jsbjx7
property-branch-nick: 5.1-innodb
testament3-sha1: 410a750cf25cf27ec78c45592962a17ebdfaf2b2