Bug#59699 - Hidden bug would cause a crash for unknown tablespace flags.
The bug would cause a crash of InnoDB if a non-standard or unknown table
flags existed in a SYS_TABLES record. This is important because the next
file version, Cheetah, will identify itself by expanding this field.  So
unless this is fixed, an older engine that tries to open a table in a
tablespace with a newer file version will crash instead of report an error
and refuse to open the table, as it should do.

Reviewed at RB://583.    Approved by Marko.
--BZR--
revision-id: kevin.lewis@oracle.com-20110204145930-jcymzrhjn0czsqww
property-branch-nick: mysql-5.5-innodb
testament3-sha1: d203f3dc3158fc1d357d41d57f62e4b2fc1b02b2