Bug #11754461 CANNOT ALTER TABLE WHEN KEY PREFIX TOO LONG
The problem was that doing ALTER TABLE on a table which had a key
on a TEXT/BLOB column with a prefix longer than the maximum number
of characteres in this column (as per the character set), by mistake,
caused an error (Error 1170 - ER_BLOB_KEY_WITHOUT_LENGTH).

This bug not repeatable in 5.5.

This patch adds a regression test to alter_table.test and
contains no code changes.

--BZR--
revision-id: jon.hauglid@oracle.com-20110221113008-s43b6dw4mj1gtg3w
property-branch-nick: mysql-5.5-bug11754461
testament3-sha1: ca38e8da0d721fbebbb3fbc9aee5862f27a8ac62