Bug#11767480 - SPATIAL INDEXES ON NON-SPATIAL COLUMNS CAUSE CRASHES.
This is a backport of the patch for MySQL Bug#50574.
      
Adding a SPATIAL INDEX on non-geometrical columns caused a
segmentation fault when the table was subsequently 
inserted into.
            
A test was added in mysql_prepare_create_table to explicitly
check whether non-geometrical columns are used in a
spatial index, and throw an error if so.
      
For MySQL 5.5 and later, a new and more meaningful error 
message was introduced.  For 5.1, we (re-)use an existing
error code.

--BZR--
revision-id: magne.mahre@oracle.com-20110224112338-dggtz05sv88otuvt
property-branch-nick: mysql-5.1-security
testament3-sha1: 4c7b24eecb96258a3106860e0c31bbee2bcf7fd6