Fix BUG#11798085 - INCORRECT INTEGER TYPES USED IN CALCULATION RESULT IN OVERFLOW
Do not assign the result of the difference to a signed variable and
checking whether it is negative afterwards because this limits the max diff
to 2G on 32 bit systems. E.g. "signed = 3.5G - 1G" would be negative and the
code would assume that 3.5G < 1G. Instead compare the two variables directly
and assign to unsigned only if we know that the result of the subtraction
will be positive.

Discussed with:	Jimmy and Sunny (via IRC)

--BZR--
revision-id: vasil.dimov@oracle.com-20110225095018-jgmv1pnuprrjzat1
property-branch-nick: mysql-5.1-innodb
testament3-sha1: 6d1453b1a435c7f3e943f64ef15bcdf950662b44