A patch for Bug#11765297 (58251 - archive_plugin and blackhole_plugin fails when running with ps-protocol).
The problem was that when running in --ps-protocol mode mysqltest.cc
didn't close created prepared statements. So, the plugins could not be
unistalled because there was a prepared statement using them.

A fix is to add a dummy statement that forces mysqltest.cc to close
the last prepared statement (which uses a plugin-defined table).
--BZR--
revision-id: alexander.nozdrin@oracle.com-20110314110308-avxfnzy61z9vtlqz
property-branch-nick: mysql-5.5-bug58251
testament3-sha1: 1bfefb727ba7d4d8765528e0785b54804e05db12