A patch for Bug#11763413 (56115: SELECT doesn't work in prepared statements with cursor protocol).
The problem was a bug in Materialized-cursor implementation.
Materialized_cursor::open() called send_result_metadata()
with items pointing to already closed table.

The fix is to send metadata when the table is still open.

NOTE: this is a "partial" fix: metadata are different with
and without --cursor-protocol, but that's a different large
problem, one indication of which is reported as Bug 24176.
--BZR--
revision-id: alexander.nozdrin@oracle.com-20110322114856-zfice2275id9s6s7
property-branch-nick: mysql-5.5-bug56115
testament3-sha1: 6b590177f373dda2ffa000994587e6a0728fb1a2