A patch for Bug#12325375: THE SERVER ON WINXP DOES NOT ALLOW CONNECTIONS IF NO DNS-SERVER AVAILABLE.
The thing is that on Windows XP getnameinfo() returns WSANO_DATA
when hostname-lookup is not available. The problem was that
this error code was treated as serious error and the client
connection got rejected.

The fix is to treat all errors from getnameinfo() as not ciritical,
but add IP-address to the host cache only for EAI_NONAME (or WSANO_DATA).
--BZR--
revision-id: alexander.nozdrin@oracle.com-20110408102336-ho3jrxz2r5nfinrj
property-branch-nick: mysql-5.5-bug12325375
testament3-sha1: da35307b3b843103fae9a0f4da6a80fdbc08e5c0