Patch for Bug#12374486 - SEVERE MEMORY LEAK IN PREPARED STATEMENTS THAT CALL STORED PROCEDURES.
The bug was introduced by WL#4435. The problem was that if a stored
procedure generated a few result sets with different set of columns,
a new memory would be allocated after every EXECUTE for every
result set.

The fix is to introduce a new memory root in scope of MYSQL_STMT,
and to store result-set metadata in that memory root.
--BZR--
revision-id: alexander.nozdrin@oracle.com-20110506133920-oplt3dt161br6z5r
property-branch-nick: mysql-5.5-bug12374486
testament3-sha1: 8be5c83a3b3cd1c272c4119ad62725be067798ee