BUG#11746302: 25228: RPL_RELAYSPACE.TEST FAILS ON POWERMACG5,               VM-WIN2003-32-A, SLES10-IA64-A
The test case waits for master_pos_wait not to timeout, which
means that the deadlock between SQL and IO threads was 
succesfully and automatically dealt with.
      
However, very rarely, master_pos_wait reports a timeout. This
happens because the time set for master_pos_wait to wait was
too small (6 seconds). On slow test env this could be a 
problem.
      
We fix this by setting the timeout inline with the one used
in sync_slave_with_master (300 seconds). In addition we 
refactored the test case and refined some comments.
--BZR--
revision-id: luis.soares@oracle.com-20110519154545-0iwgrq2tugv8zo8q
property-branch-nick: mysql-5.1-gca
testament3-sha1: d1314378665da382d04845787183904907c1288d