Bug#12563279: REGRESSION IN HANDLING PRE-4.1 AUTHENTICATION PACKET
The problem is that clients implementing the 4.0 version of the
protocol (that is, mysql-4.0) do not null terminate a string
at the end of the authentication packet. These clients denote
the end of the string with the end of the packet.

Although this goes against the documented (see MySQL Internals
ClientServer Protocol wiki) description of the protocol, these
old clients still need to be supported.

The solution is to support the documented and actual behavior
of the clients. If a client is using the pre-4.1 version of
the protocol, the end of a string in the authentication packet
can either be denoted with a null character or by the end of
the packet. This restores backwards compatibility with old
clients implementing either the documented or actual behavior.
--BZR--
revision-id: davi.arnaut@oracle.com-20110530104230-0glodzdalw635ulv
property-branch-nick: 12563279-5.1
property-file-info: ld7:file_id63:sp1f-password.c-19700101030959-hyqrqz6vp4g26qbqvd4fq6z2ecdiqerx7:message297:The scrambled message, as provided by the user, might not be
property-file-info: properly null terminated. If this is the case, uninitialized
property-file-info: memory past the end of the buffer could theoretically be
property-file-info: accessed. To ensure that this is never the case, copy the
property-file-info: scrambled message over to a null terminated auxiliar buffer.4:path14:sql/password.ced7:file_id67:sp1f-sql_connect.cc-20070223111352-fhh5znxdfvzxuca7da3uu4olnwgkrm4n7:message449:Use different execution paths to read strings depending on the
property-file-info: protocol being used. If version 4.0 of the protocol is used,
property-file-info: end of string can be denoted with a NUL character or by the
property-file-info: end of the packet.
property-file-info: 
property-file-info: If there are not enough bytes left after the current position
property-file-info: of the buffer to satisfy the current string, the string is
property-file-info: considered to be empty. This is required because old clients
property-file-info: do not send the password string field if the password is empty.4:path18:sql/sql_connect.ccee
testament3-sha1: 4faee633ad6533a43255528a5aa5ba57447a900d